requirements.txt
alembic==1.18.4
Authlib==1.7.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.3
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.4.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joserfc==1.6.4
Mako==1.3.11
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
pycparser==3.0
python-dotenv==1.2.2
requests==2.33.1
SQLAlchemy==2.0.49
SQLAlchemy-Utils==0.42.1
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
WTForms==3.2.1
